backend/requirements.txt
fastapi
uvicorn[standard]
groq
pydantic-settings
python-dotenv
python-pptx
reportlab
